\name{Aquila-package}
\alias{Aquila-package}
\alias{Aquila}
\docType{package}
\title{
\packageTitle{Aquila}
}
\description{
\packageDescription{Aquila}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Aquila}
\packageIndices{Aquila}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Aquila}

Maintainer: \packageMaintainer{Aquila}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
