% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetSignacObject}
\alias{GetSignacObject}
\title{Create Signac Object}
\usage{
GetSignacObject(mrna.h5.file, peaks.fragments.file)
}
\arguments{
\item{mrna.h5.file}{Path to the input mrna h5 file}

\item{peaks.fragments.file}{Path to the input peaks file}
}
\value{
A Signac object
}
\description{
This function loads the mRNA and peaks file as a Signac object.
}
